refactor: business logic 경로 정리
