+++ src/business/constants.ts
@@ -1,0 +1,5 @@
+export const LOTTO_PRICE = 1_000;
+
+export const LOTTO_NUMBERS_LENGTH = 6;
+
+export const LOTTO_MAX_NUMBER = 45;
