+++ src/business/domain/LottoTicket.ts
@@ -1,0 +1,10 @@
+import { LottoNumberGenerator } from "../generator/LottoNumberGenerator";
+import { LottoNumber } from "./LottoNumber";
+
+export class LottoTicket {
+  readonly lottoNumbers: LottoNumber[];
+
+  constructor() {
+    this.lottoNumbers = LottoNumberGenerator.generate();
+  }
+}
