+++ src/business/enum.ts
@@ -1,0 +1,15 @@
+export enum PRIZE_NAME {
+  FIRST = "FIRST",
+  SECOND = "SECOND",
+  THIRD = "THIRD",
+  FOURTH = "FOURTH",
+  FAIL = "FAIL",
+}
+
+export enum PRIZE_MONEY {
+  FIRST = 2_000_000_000,
+  SECOND = 1_500_000,
+  THIRD = 50_000,
+  FOURTH = 5_000,
+  FAIL = 0,
+}
