+++ src/view/components/PurchaseMoney.ts
@@ -4,6 +4,7 @@ import { Input } from "./common/Input";
 export class PurchaseMoney {
   private readonly input: Input;
   private readonly confirmButton: Button;
+  readonly money: number | null;
 
   constructor({
     input,
@@ -14,5 +15,6 @@ export class PurchaseMoney {
   }) {
     this.input = input;
     this.confirmButton = confirmButton;
+    this.money = null;
   }
 }
